refactor: StudyRecruitPage 코드 컴포넌트 분리 #7
