+++ src/pages/StudyRecruitPage.tsx
@@ -1,11 +1,8 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import styled from "styled-components";
 import StarIcon from "../assets/icons/StarIcon.png";
-import DaumPostcode from "react-daum-postcode";
-import { useNavigate } from "react-router-dom";
-import { LatLng } from "../types";
-import { PostRequest } from "../types/api/post";
-import { usePostCreate } from "../hooks/api/usePost";
+import { AddressSearch } from "../components/AddressSearch";
+import { useRecruitForm } from "../hooks/study/useRecruitForm";
 
 const PageContainer = styled.div`
   padding: 50px 100px 50px 100px;
@@ -107,218 +104,22 @@ const RowWrapper = styled.div`
   align-items: center;
 `;
 
-const AddressInput = styled.input`
-  margin-top: 5px;
-  margin-bottom: 15px;
-  padding: 15px;
-  border: 3px solid #b3b4dc;
-  border-radius: 10px;
-  width: 28vw;
-  font-size: 18px;
-  font-family: "SCDream4", sans-serif;
-`;
-
-const FindAddressButton = styled.button`
-  margin-left: 20px;
-  width: 100px;
-  height: 50px;
-  border: none;
-  border-radius: 10px;
-  background-color: #b6b6b6;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
-  font-family: "SCDream6";
-  color: white;
-  font-size: 16px;
-  cursor: pointer;
-`;
-
-interface AddressObj {
-  areaAddress: string;
-  townAddress: string;
-}
-
-interface FindAddressProps {
-  setAddressObj: React.Dispatch<React.SetStateAction<AddressObj>>;
-  setLatLng: React.Dispatch<React.SetStateAction<LatLng>>;
-  setFormData: React.Dispatch<React.SetStateAction<PostRequest>>;
-}
-
-interface DaumPostcodeData {
-  address: string;
-  addressType: string;
-  bname: string;
-  buildingName: string;
-}
-
-function FindAddress({ setAddressObj, setLatLng, setFormData }: FindAddressProps) {
-  const handleComplete = useCallback((data: DaumPostcodeData) => {
-    // 도로명 주소의 노출 규칙에 따라 주소를 표시
-    let fullAddress = data.address;
-    let extraAddress = "";
-    if (data.addressType === "R") {
-      if (data.bname !== "") {
-        extraAddress += data.bname;
-      }
-      if (data.buildingName !== "") {
-        extraAddress +=
-          extraAddress !== "" ? `, ${data.buildingName}` : data.buildingName;
-      }
-
-      fullAddress += extraAddress !== "" ? `(${extraAddress})` : "";
-
-      setAddressObj({
-        areaAddress: "",
-        townAddress: fullAddress,
-      });
-
-      setFormData(prev => ({
-        ...prev,
-        place: fullAddress
-      }));
-
-      // 사용자가 입력한 주소 정보를 입력 필드에 넣음
-      const addressInput = document.getElementById("addressInput") as HTMLInputElement;
-      if (addressInput) {
-        addressInput.value = fullAddress;
-      }
-
-      // 주소로 좌표를 검색 (Kakao Map API 사용)
-      window.kakao.maps.load(() => {
-        const geocoder = new window.kakao.maps.services.Geocoder();
-        geocoder.addressSearch(fullAddress, (result: any, status: any) => {
-          if (status === window.kakao.maps.services.Status.OK) {
-            const latitude = parseFloat(result[0].y);
-            const longitude = parseFloat(result[0].x);
-
-            setAddressObj({
-              areaAddress: "",
-              townAddress: fullAddress,
-            });
-
-            // 입력받은 주소의 위도, 경도 정보를 state에 저장
-            setLatLng({
-              latitude,
-              longitude,
-            })
-
-            const addressInput = document.getElementById("addressInput") as HTMLInputElement;
-            if (addressInput) {
-              addressInput.value = fullAddress;
-            }
-          }
-        });
-      });
-    }
-  }, []);
-
-  // 주소 검색 API를 이용해 주소 찾기
-  useEffect(() => {
-    const script = document.createElement("script");
-    script.type = "text/javascript";
-    script.src = `//dapi.kakao.com/v2/maps/sdk.js?appkey=${process.env.REACT_APP_KAKAOMAP_API_KEY}&libraries=services`;
-    document.head.appendChild(script);
-
-    // Kakao Maps API가 로드되면 실행
-    script.onload = () => {
-      if (window.kakao?.maps?.services && window?.daum) {
-        const geocoder = new window.kakao.maps.services.Geocoder();
-        const address = new window.daum.Postcode({
-          oncomplete: handleComplete,
-        });
-        // @ts-ignore
-        window.address = address;
-      } else {
-        console.error("Failed to load Kakao Maps API");
-      }
-    };
-
-    script.onerror = () => {
-      console.error("Failed to load Kakao Maps SDK script");
-    };
-
-    return () => {
-      document.head.removeChild(script);
-    };
-  }, [handleComplete]);
-
-  return (
-    // @ts-ignore
-    <FindAddressButton type="button" onClick={() => window.address.open()}>
-      주소 찾기
-    </FindAddressButton>
-  );
-}
-
 const StudyRecruitPage = () => {
-  const [addressObj, setAddressObj] = useState<AddressObj>({
-    areaAddress: "",
-    townAddress: "",
-  });
-
-  const [latLng, setLatLng] = useState<LatLng>({
-    latitude: null,
-    longitude: null,
-  });
-
-  const userInfo = JSON.parse(localStorage.getItem('userInfo') || 'null');
-
-  const [formData, setFormData] = useState<PostRequest>({
-    userId: userInfo?.userId || 0,
-    skill: "",
-    place: "",
-    latitude: 0,
-    longitude: 0,
-    progress: "",
-    peopleNum: 0,
-    deadline: "",
-    type: "",
-    done: false,
-    title: "",
-    content: ""
-  });
-
-  const navigate = useNavigate();
-  const postCreate = usePostCreate();
-
-  const handleSubmit = async () => {
-    try {
-      if (!latLng.latitude || !latLng.longitude) {
-        alert("주소를 선택해주세요.");
-        return;
-      }
-      
-      const postData: PostRequest = {
-        ...formData,
-        latitude: latLng.latitude,
-        longitude: latLng.longitude,
-        place: addressObj.townAddress,
-      }
-
-      await postCreate.mutateAsync(postData);
-
-      alert("글이 등록되었습니다.");
-      navigate("/");
-    } catch (error) {
-      alert("글 등록에 실패했습니다.");
-      console.error("Error:", error);
-    }
-  };
+  const {
+    formData,
+    latLng,
+    setFormData,
+    setAddressObj,
+    setLatLng,
+    handleInputChange,
+    handleSubmit,
+  } = useRecruitForm();
 
   useEffect(() => {
     console.log("Latitude:", latLng.latitude);
     console.log("Longitude:", latLng.longitude);
     // window.location.reload();
   }, [latLng]);
-
-  const handleInputChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    const { name, value } = e.target;
-    setFormData(prev => ({
-      ...prev,
-      [name]: name === 'peopleNum' ? Number(value) : value
-    }));
-  };
 
   return (
     <PageContainer>
@@ -379,18 +180,12 @@ const StudyRecruitPage = () => {
         <Inputbox>
           <InputWrapper>
             <TextInput>진행 장소</TextInput>
-            <AddressInput
-              type="text"
-              id="addressInput"
-              name="place"
-              placeholder="주소를 입력해주세요."
-              value={formData.place}
-              onChange={handleInputChange}
-            />
-            <FindAddress
+            <AddressSearch
+              addressValue={formData.place}
               setAddressObj={setAddressObj}
               setLatLng={setLatLng}
               setFormData={setFormData}
+              handleInputChange={handleInputChange}
             />
           </InputWrapper>
           <InputWrapper>
